Fonctionnalité pour update le delta et fix de l'affichage cryptocurrencies list

+++ src/main/java/com/cda/jee/crypto/model/CryptoCurrency.java
@@ -17,7 +17,7 @@ public class CryptoCurrency {
 	private int idCrypto;
 	private String name;
 	private String symbol;
-	private int delta;
+	private double delta;
 	private double currentPrice;
 	private String imageUrl;
 	private LocalDateTime lastUpdated;
